+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Gunpowder Teleport 1.0.2
+
+- Update to Gunpowder 0.3.1 (1.16.4)
+
+---
 # Gunpowder Teleport 1.0.1
 
 - /home, /spawn and /warp now teleports to the center of the block